alterações main

+++ modulos/bruteForce.h
@@ -3,6 +3,6 @@
 
 int max(int a, int b);
 
-unsigned long int bruteForce(unsigned long int *vet, int n, int i);
+unsigned long int bruteForce(unsigned long int *vet, unsigned long int n, int i);
 
 #endif
